Fixed offset_u

+++ flecsi/utils/offset.h
@@ -41,8 +41,13 @@ public:
    *
    * Since we ensure that COUNT_BITS <= 32, this will always fit into a
    * uint32_t.
+   *
+   * While this is true, the compliment of a uint32_t is much different
+   * than the compliment of a uint64_t.  Thus for safety and simplicity,
+   * we store count_max as a uint64_t for simplicity.  This will avoid having
+   * to statically cast count_mask.
    */
-  static constexpr uint32_t count_mask = (1ul << COUNT_BITS) - 1;
+  static constexpr uint64_t count_mask = (1ul << COUNT_BITS) - 1;
 
   /**
    * @brief The maximum value of the start index value.
